treasury to module
