+++ frontend/src/store/store.ts
@@ -35,7 +35,6 @@ import {getWeaponNameFromSeed} from '@/weapon-name';
 import axios from 'axios';
 import {abi as ierc20Abi} from '@/../../build/contracts/IERC20.json';
 import {abi as erc20Abi} from '@/../../build/contracts/ERC20.json';
-import {SupportedProject} from '@/views/Treasury.vue';
 import {abi as erc721Abi} from '@/../../build/contracts/IERC721.json';
 import BigNumber from 'bignumber.js';
 
@@ -68,6 +67,7 @@ import quests from './quests';
 import raid from './raid';
 import staking from './staking';
 import land from './land';
+import treasury from './treasury';
 
 Vue.use(Vuex);
 
@@ -78,7 +78,8 @@ export default new Vuex.Store<IState>({
     quests,
     raid,
     staking,
-    land
+    land,
+    treasury
   },
   state: {
     web3: new Web3(),
@@ -3443,172 +3444,6 @@ export default new Vuex.Store<IState>({
 
       await Promise.all([
         dispatch('fetchCharacterCosmetic', id)
-      ]);
-    },
-
-    async addPartnerProject({state}, {partnerProject}) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      return await Treasury.methods.addPartnerProject(
-        partnerProject.name,
-        partnerProject.tokenSymbol,
-        partnerProject.tokenAddress,
-        partnerProject.tokenSupply,
-        Web3.utils.toWei(partnerProject.tokenPrice.toString(), 'ether').toString(),
-        partnerProject.distributionTime,
-        partnerProject.isActive,
-        partnerProject.logo,
-        partnerProject.details,
-        partnerProject.website,
-        partnerProject.note,
-      ).send({from: state.defaultAccount, gasPrice: getGasPrice()});
-    },
-
-    async getActivePartnerProjects({state}) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      const ids = await Treasury.methods.getActivePartnerProjectsIds().call(defaultCallOptions(state));
-      const projects = [];
-      for(let i = 0; i < ids.length; i++) {
-        const project = await Treasury.methods.partneredProjects(ids[i]).call(defaultCallOptions(state));
-        projects.push(project);
-      }
-      return projects;
-    },
-
-    async setPartnerProjectLogo({state}, {id, logo}) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      await Treasury.methods.setProjectLogo(id, logo).send({from: state.defaultAccount, gasPrice: getGasPrice()});
-    },
-
-    async setPartnerProjectDetails({state}, {id, details}) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      await Treasury.methods.setProjectDetails(id, details).send({from: state.defaultAccount, gasPrice: getGasPrice()});
-    },
-
-    async setPartnerProjectWebsite({state}, {id, website}) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      await Treasury.methods.setProjectWebsite(id, website).send({from: state.defaultAccount, gasPrice: getGasPrice()});
-    },
-
-    async setPartnerProjectNote({state}, {id, note}) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      await Treasury.methods.setProjectNote(id, note).send({from: state.defaultAccount, gasPrice: getGasPrice()});
-    },
-
-    async setPartnerProjectIsActive({state}, {id, isActive}) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      await Treasury.methods.setIsActive(id, isActive).send({from: state.defaultAccount, gasPrice: getGasPrice()});
-    },
-
-    async fetchPartnerProjects({ state, dispatch }) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      const activePartnerProjectIds = await Treasury.methods.getActivePartnerProjectsIds().call(defaultCallOptions(state));
-      activePartnerProjectIds.forEach(async (id: string) => {
-        await dispatch('fetchPartnerProject', id);
-      });
-
-      await dispatch('fetchDefaultSlippage');
-    },
-
-    async fetchPartnerProject({ state, commit }, id) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      const partnerProjectRaw = await Treasury.methods.partneredProjects(id).call(defaultCallOptions(state));
-      const tokensClaimed = await Treasury.methods.tokensClaimed(id).call(defaultCallOptions(state));
-      const data = await Treasury.methods.getProjectData(id).call(defaultCallOptions(state));
-
-      const partnerProject = {
-        id: +partnerProjectRaw[0],
-        name: partnerProjectRaw[1],
-        tokenSymbol: partnerProjectRaw[2],
-        tokenAddress: partnerProjectRaw[3],
-        tokenSupply: +partnerProjectRaw[4],
-        tokensClaimed: +tokensClaimed,
-        tokenPrice: +partnerProjectRaw[5],
-        isActive: partnerProjectRaw[6],
-        logo: data[0],
-        details: data[1],
-        website: data[2],
-        note: data[3],
-      } as SupportedProject;
-
-      commit('updatePartnerProjectsState', { partnerProjectId: partnerProject.id, partnerProject });
-    },
-
-    async fetchDefaultSlippage({ state, commit }) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      const slippage = await Treasury.methods.defaultSlippage().call(defaultCallOptions(state));
-
-      commit('updateDefaultSlippage', slippage);
-    },
-
-    async getPartnerProjectMultiplier({ state, commit }, id) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      const multiplier = await Treasury.methods.getProjectMultiplier(id).call(defaultCallOptions(state));
-      commit('updatePartnerProjectMultiplier', { partnerProjectId: id, multiplier });
-
-      return multiplier;
-    },
-
-    async getPartnerProjectDistributionTime({ state }, id) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      return await Treasury.methods.projectDistributionTime(id).call(defaultCallOptions(state));
-    },
-
-    async getPartnerProjectClaimedAmount({ state }, id) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      return await Treasury.methods.tokensClaimed(id).call(defaultCallOptions(state));
-    },
-
-    async getSkillToPartnerRatio({ state, commit }, id) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      const ratio = await Treasury.methods.getSkillToPartnerRatio(id).call(defaultCallOptions(state));
-      commit('updatePartnerProjectRatio', { partnerProjectId: id, ratio });
-
-      return ratio;
-    },
-
-    async claimPartnerToken({ state, dispatch },
-                            { id, skillAmount, currentMultiplier, slippage }:
-                            {id: number, skillAmount: string, currentMultiplier: string, slippage: string}) {
-      const { Treasury } = state.contracts();
-      if(!Treasury || !state.defaultAccount) return;
-
-      await Treasury.methods.claim(id, skillAmount, currentMultiplier, slippage).send({
-        from: state.defaultAccount,
-        gasPrice: getGasPrice()
-      });
-
-      await Promise.all([
-        dispatch('fetchSkillBalance'),
-        dispatch('fetchFightRewardSkill'),
-        dispatch('fetchPartnerProject', id)
       ]);
     },
